% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demogRTF.R
\name{demogRTF_server}
\alias{demogRTF_server}
\title{Demographics Table RTF - UI}
\usage{
demogRTF_server(input, output, session, params)
}
\arguments{
\item{input}{module input}

\item{output}{module output}

\item{session}{module session}

\item{params}{parameters object with \code{data} and \code{settings} options.}
}
\value{
returns shiny module Server function
}
\description{
Demographics Table RTF - UI
}
