% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demogRTF.R
\name{demogRTF_table}
\alias{demogRTF_table}
\title{create demographics RTF table}
\usage{
demogRTF_table(data, settings)
}
\arguments{
\item{data}{demographics data frame with columns specified in settings object}

\item{settings}{list with parameters specifying the column names for:
\itemize{
\item sex (settings$sex_col),
\item race (settings$race_col)
\item age (settings$age_Col)
}}
}
\value{
rtf doc object
}
\description{
create demographics RTF table
}
\examples{
dm <- read.csv("https://raw.githubusercontent.com/RhoInc/data-library/master/data/clinical-trials/sdtm/cdisc-pilot-01/dm.csv")
settings <- list(treatment_col = "ARM", sex_col = "SEX", race_col = "RACE", age_col = "AGE")
demogRTF_table(data, settings)

}
