% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_demogRTF.R
\name{demogRTF_ui}
\alias{demogRTF_ui}
\title{Demographics Table RTF - UI}
\usage{
demogRTF_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
returns shiny module UI
}
\description{
Demographics Table RTF - UI
}
