% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_aeExplorer.R
\name{init_aeExplorer}
\alias{init_aeExplorer}
\title{Initialize Settings for Adverse Event Explorer widget}
\usage{
init_aeExplorer(data, settings)
}
\arguments{
\item{data}{adverse events data structured as one record per event. See details for column requirements.}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for Adverse Event Explorer widget
}
