% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_aeTimeline.R
\name{init_aeTimeline}
\alias{init_aeTimeline}
\title{Initialize Settings for Adverse Event Timeline}
\usage{
init_aeTimeline(data, settings)
}
\arguments{
\item{data}{labs data structured as one record per person per visit per measurement. See details for column requirements.}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for Adverse Event Timeline
}
