% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_aeTimelines.R
\name{init_aeTimelines}
\alias{init_aeTimelines}
\title{Initialize Settings for AE Timeline widget}
\usage{
init_aeTimelines(data, settings)
}
\arguments{
\item{data}{adverse events data structured as one record per event. See details for column requirements.}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for AE Timeline widget
}
