% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_mhExplorer.R
\name{init_mhExplorer}
\alias{init_mhExplorer}
\title{Initialize Settings for Medical History Explorer widget}
\usage{
init_mhExplorer(data, settings)
}
\arguments{
\item{data}{medical history and demographics data.}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for Medical History Explorer widget
}
