% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_paneledOutlierExplorer.R
\name{init_paneledOutlierExplorer}
\alias{init_paneledOutlierExplorer}
\title{Initialize Settings for Paneled Outlier Explorer widget}
\usage{
init_paneledOutlierExplorer(data, settings)
}
\arguments{
\item{data}{labs data structured as one record per person per visit per measurement. See details for column requirements.}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for Paneled Outlier Explorer widget
}
