% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_safetyOutlierExplorer.R
\name{init_safetyOutlierExplorer}
\alias{init_safetyOutlierExplorer}
\title{Initialize Settings for Safety Outlier Explorer widget}
\usage{
init_safetyOutlierExplorer(data, settings)
}
\arguments{
\item{data}{labs data structured as one record per person per visit per measurement. See details for column requirements.}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for Safety Outlier Explorer widget
}
