% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_safetyResultsOverTime.R
\name{init_safetyResultsOverTime}
\alias{init_safetyResultsOverTime}
\title{Initialize Settings for Safety Results Over Time widget}
\usage{
init_safetyResultsOverTime(data, settings)
}
\arguments{
\item{data}{labs data structured as one record per person per visit per measurement. See details for column requirements.}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for Safety Results Over Time widget
}
