% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_safetyShiftPlot.R
\name{init_safetyShiftPlot}
\alias{init_safetyShiftPlot}
\title{Initialize Settings for Safety Shift Plot widget}
\usage{
init_safetyShiftPlot(data, settings)
}
\arguments{
\item{data}{labs data structured as one record per person per visit per measurement. See details for column requirements.}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for Safety Shift Plot widget
}
