% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lab_distribution.R
\name{lab_distribution_server}
\alias{lab_distribution_server}
\title{lab distribution Module - Server}
\usage{
lab_distribution_server(input, output, session, params)
}
\arguments{
\item{input}{module input}

\item{output}{module output}

\item{session}{module session}

\item{params}{parameters object with \code{data} and \code{settings} options.}
}
\value{
returns shiny module Server function
}
\description{
A simple server for a shiny module looking at lab histograms. Intended primarily for technical demos.
}
