% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lab_distribution.R
\name{lab_distribution_ui}
\alias{lab_distribution_ui}
\title{Lab distribution Module - UI}
\usage{
lab_distribution_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
returns shiny module UI
}
\description{
A simple UI for a shiny module looking at lab histograms. Intended primarily for technical demos.
}
