% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{meta_timeToAE}
\alias{meta_timeToAE}
\title{Metadata data frame containing information about the data mapping used to configure safetyGraphics for the Time to AE Chart. One record per unique data mapping}
\format{
A data frame with X rows and 10 columns
\describe{
\item{domain}{Data domain}
\item{text_key}{Text key indicating the setting name. \code{'--'} delimiter indicates a field level data mapping}
\item{col_key}{Key for the column mapping}
\item{field_key}{Key for the field mapping (if any)}
\item{type}{type of mapping - "field" or "column"}
\item{label}{Label}
\item{description}{Description}
\item{multiple}{Mapping supports multiple columns/fields }
\item{standard_adam}{Default values for the ADaM data standard}
\item{standard_sdtm}{Default values for the SDTM data standard}
}
}
\source{
Created for this package
}
\usage{
meta_timeToAE
}
\description{
Metadata data frame containing information about the data mapping used to configure safetyGraphics for the Time to AE Chart. One record per unique data mapping
}
\keyword{datasets}
