% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paneledOutlierExplorer.R
\name{paneledOutlierExplorer}
\alias{paneledOutlierExplorer}
\title{Make standalone paneledOutlierExplorer html widget}
\usage{
paneledOutlierExplorer(df = safetyData::adam_adlbc, mapping = NULL, ...)
}
\arguments{
\item{df}{data frame containing lab data used to render for paneledOutlierExplorer. Default is safetyData::adam_adlbc.}

\item{mapping}{named list with the current data mappings. See details for default mapping.}

\item{...}{additional options to be added to mapping. Will overwrite mapping.}
}
\value{
An \code{htmlwidget} object of class \code{paneledOutlierExplorer}, ready to
print in the RStudio viewer, embed in R Markdown, or render inside a Shiny
application.
}
\description{
Convience mapping of render_widget for paneledOutlierExplorer.
}
\details{
The \href{https://github.com/Rhoinc/paneled-outlier-explorer/wiki/Data-Guidelines}{data} and \href{https://github.com/RhoInc/paneled-outlier-explorer/wiki/Configuration}{mapping} should match the specs described in the \href{https://github.com/rhoinc/paneled-outlier-explorer/wiki/Configuration}{paneledOutlierExplorer} javascript library. Items passed in ... are added to mapping, and then the list is converted to json via \code{jsonlite::toJSON(mapping, auto_unbox=TRUE, null="null")}.

The default mapping shown below is designed to work with data in the CDISC ADaM format (like \code{safetydata::adam_adlbc}).

\if{html}{\out{<div class="sourceCode">}}\preformatted{mapping <-  list(
       measure_col = 'PARAM',
       time_cols = list(
         list(
           value_col = "VISIT",
           order_col = "VISITNUM",
           label = "VISIT",
           type = 'ordinal',
           label="Visit",
           rotate_tick_labels = TRUE,
           vertical_space = 75
         ),
         list(
           value_col = "ADY",
           order_col = "ADY",
           label = "Study Day",
           type = 'linear',
           label="Visit",
           rotate_tick_labels = FALSE,
           vertical_space = 0
         )
       ),
       value_col = 'AVAL',
       id_col = 'USUBJID',
       lln_col = 'A1LO',
       uln_col = 'A1HI'
     )                
}\if{html}{\out{</div>}}

Parameters that are not included in the default mapping can be accessed via ...; Key options and defaults for safetyData::adam_adlbc shown below:
\itemize{
\item \code{filters}: list of columns to be included as data filters formatted as `filters=c("SEX","AGEGR1")``)
}

For more options see the \href{https://github.com/rhoinc/paneled-outlier-explorer/wiki/Configuration}{full specs} in the javascript library.
}
\examples{
\dontshow{if (requireNamespace("safetyData", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Render widget with defaults
paneledOutlierExplorer() 

# Add Sex and Age Filters
paneledOutlierExplorer(
    filters=list(
        list(value_col="SEX"),
        list(value_col="AGEGR1",label="Age")
    )
)

# customize panel size (in pixels)
paneledOutlierExplorer(multiples_sizing=list("width"= 500,"height"= 300))

# customize default normal range
paneledOutlierExplorer(
   normal_range_method = 'quantiles',
   normal_range_quantile_low = 0.2,
   normal_range_quantile_high = 0.8,
)
\dontshow{\}) # examplesIf}
}
