% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_QT_Explorer.R
\name{QT_Explorer_server}
\alias{QT_Explorer_server}
\title{QT Explorer Module - UI}
\usage{
QT_Explorer_server(input, output, session, params)
}
\arguments{
\item{input}{module input}

\item{output}{module output}

\item{session}{module session}

\item{params}{parameters object with \code{data} and \code{settings} options.}
}
\value{
returns shiny module Server function
}
\description{
QT Explorer Module - UI
}
