% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_QT_Explorer.R
\name{QT_Explorer_ui}
\alias{QT_Explorer_ui}
\title{QT Explorer Module - UI}
\usage{
QT_Explorer_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
returns shiny module UI
}
\description{
QT Explorer Module - UI
}
