% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_QT_OutlierExplorer.R
\name{QT_OutlierExplorer_ui}
\alias{QT_OutlierExplorer_ui}
\title{QT Outlier Explorer Module - UI}
\usage{
QT_OutlierExplorer_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
returns shiny module UI
}
\description{
QT Outlier Explorer Module - UI
}
