% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_histogram.R
\name{safety_histogram}
\alias{safety_histogram}
\title{Safety Histogram}
\usage{
safety_histogram(data, settings, description = "")
}
\arguments{
\item{data}{labs data set}

\item{settings}{list of settings including the following parameters:}

\item{description}{Description of the chart to be printed as a subtitle}
}
\value{

}
\description{
Safety Histogram
}
