% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_results_over_time.R
\name{safety_results_over_time}
\alias{safety_results_over_time}
\title{Safety Results Over Time plot}
\usage{
safety_results_over_time(data, settings, description)
}
\arguments{
\item{data}{}

\item{settings}{}

\item{description}{}
}
\value{
plot object
}
\description{
Safety Results Over Time plot
}
