% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_safetyOutlierExplorer.R
\name{safetyOutlierExplorer_ui}
\alias{safetyOutlierExplorer_ui}
\title{Safety Outlier Explorer Module - UI}
\usage{
safetyOutlierExplorer_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
returns shiny module UI
}
\description{
Safety Outlier Explorer Module - UI
}
