% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_events.R
\name{standardize_events}
\alias{standardize_events}
\title{Create a standardized event data set}
\usage{
standardize_events(data, settings)
}
\arguments{
\item{data}{safetyGraphics data object}

\item{settings}{safetyGraphics settings object.}

\item{domains}{list of safetyGraphics domains. Adverse events ('aes'), concominate medications ('cm') and exposure ('ex') included by default.}
}
\value{
combined dataset with stacked AE and CM data
}
\description{
Create an event data set with a standard set of hard-coded column names using standard safetyGraphics settings and data. The settings for each specified domain should contain valid mappings for ID ("id_col"), event start date ("stdy_col") and event end date ("endy_col"). Missing start day and end day values are extrapolated to NA. All other columns specified in settings are collapsed into a single "details" column. The final standardized data contains the following columns: "id", "domain", "stdy", "endy", "details".
}
