% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToAE.R
\name{timeToAE}
\alias{timeToAE}
\title{Time to AE}
\usage{
timeToAE(data, settings, ae = "Any AE", stratum = "None")
}
\arguments{
\item{data}{list of data frames including dataframes named \code{aes} (adverse events) and \code{dm} (demographics)}

\item{settings}{named list of domain-specific settings with the parameters specified below.}

\item{ae}{AE of interest. Default="Any ae"}

\item{stratum}{Default='None'}
}
\value{
returns a chart object
}
\description{
Time to AE
}
