% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tplyr_demog_chart.R
\name{tplyr_demog_chart}
\alias{tplyr_demog_chart}
\title{Demographics chart created with tplyr}
\usage{
tplyr_demog_chart(data, settings)
}
\arguments{
\item{data}{demographics data frame with columns specified in settings object}

\item{settings}{list with parameters specifying the column names for:
\itemize{
\item sex (settings$sex_col),
\item race (settings$race_col)
\item age (settings$age_Col)
}}
}
\value{
tplyr table object
}
\description{
Demographics chart created with tplyr
}
\examples{
dm <- read.csv("https://raw.githubusercontent.com/RhoInc/data-library/master/data/clinical-trials/sdtm/cdisc-pilot-01/dm.csv")
settings<-list(treatment_col="ARM", sex_col="SEX", race_col="RACE", age_col="AGE")
tplyr_demog_chart(data, settings)

}
