% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tplyr_shift_chart.R
\name{tplyr_shift_chart}
\alias{tplyr_shift_chart}
\title{Demographics chart created with tplyr}
\usage{
tplyr_shift_chart(data, settings)
}
\arguments{
\item{data}{demographics data frame with columns specified in settings object}

\item{settings}{list with parameters specifying the column names for:
\itemize{
\item ANRIND
\item BNRIND
\item TRTA
\item PARAMCD
\item VISIT
}}
}
\description{
Demographics chart created with tplyr
}
