% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_ae.R
\name{upset_ae}
\alias{upset_ae}
\title{upset AE co-occurence plot}
\usage{
upset_ae(data, settings)
}
\arguments{
\item{data}{ae data with one record per event}

\item{settings}{named list of domain-specific settings with the parameters specified below.}
}
\value{
returns a chart object
}
\description{
Create a plot using the {{upset}} package
}
\details{
The settings object provides details regarding the columns in the data sets.

\itemize{
\item{"settings$dm$id_col"}{ID column}
\item{"settings$aes$bodsys_col"}{Body System}
}
}
\examples{
aes <- safetyData::adam_adae
settings <- list(
   id_col = "USUBJID",
   bodsys_col = "AEBODSYS"
)

upset_ae(aes, settings)
}
