% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeExplorer.R
\name{aeExplorer}
\alias{aeExplorer}
\title{Make standalone aeExplorer html widget}
\usage{
aeExplorer(
  data = list(dm = safetyData::sdtm_dm, aes = safetyData::sdtm_ae),
  mapping = NULL,
  ...
)
}
\arguments{
\item{mapping}{named list with the current data mappings. See details for default mapping.}

\item{...}{additional options to be added to mapping. Will overwrite mapping.}

\item{df}{data frame containing lab data used to render for aeExplorer. Default is safetyData::adam_adlbc.}
}
\description{
Convience mapping of render_widget for aeExplorer.
}
\details{
The \href{https://github.com/RhoInc/aeexplorer/wiki/Data-Guidelines}{data} and
\href{https://github.com/RhoInc/aeexplorer/wiki/Configuration}{mapping} should match the specs
described in the \href{https://github.com/RhoInc/aeexplorer/wiki/Configuration}{aeExplorer}
JavaScript library. Items passed in ... are added to mapping, and then the list is converted to
JSON via \code{jsonlite::toJSON(mapping, auto_unbox=TRUE, NULL="NULL")}.

The default mapping shown below is designed to work with data in the CDISC ADaM format (like
\code{safetydata::adam_adlbc}).

\if{html}{\out{<div class="sourceCode">}}\preformatted{mapping <- list(
    dm = list(
        id_col = 'USUBJID',
        treatment_col = 'ARM'
    ),
    aes = list(
        id_col = 'USUBJID',
        bodsys_col = 'AEBODSYS',
        term_col = 'AEDECOD'
    )
)
}\if{html}{\out{</div>}}

Parameters that are not included in the default mapping can be accessed via ....

For more options see the \href{https://github.com/RhoInc/aeexplorer/wiki/Configuration}{full specs} in
the JavaScript library.
}
\examples{
\dontrun{
# Render widget with defaults.
aeExplorer() 

# Render widget without stratification.
aeExplorer(
    mapping = list(
        dm = list(
            id_col = 'USUBJID'
        ),
        aes = list(
            id_col = 'USUBJID',
            bodsys_col = 'AEBODSYS',
            term_col = 'AEDECOD'
        )
    )
)
}

}
