% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_aeExplorer.R
\name{init_aeExplorer}
\alias{init_aeExplorer}
\title{Initialize Settings for Adverse Event Explorer widget}
\usage{
init_aeExplorer(data, settings)
}
\arguments{
\item{data}{\code{list} Named list of data frames that includes participant-level subject data (\code{dm})
and event-level adverse event data (\code{aes}).}

\item{settings}{\code{list} Named list of settings.}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for Adverse Event Explorer widget
}
