% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_cmExplorer.R
\name{init_cmExplorer}
\alias{init_cmExplorer}
\title{Initialize Settings for Conmeds Explorer widget}
\usage{
init_cmExplorer(data, settings)
}
\arguments{
\item{data}{Con meds and demographics data. See details for column requirements.}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings
}
\description{
Initialize Settings for Conmeds Explorer widget
}
