% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_widget.R
\name{render_widget}
\alias{render_widget}
\title{Render an htmlwidget using standard safetyGraphics workflow}
\usage{
render_widget(widgetName, data, mapping)
}
\arguments{
\item{widgetName}{name of the widget saved in safetyCharts}

\item{data}{named list of current data sets}

\item{mapping}{named list with the current data mappings}
}
\description{
Render an htmlwidget using standard safetyGraphics workflow
}
