% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_outlier_explorer.R
\name{safety_outlier_explorer}
\alias{safety_outlier_explorer}
\title{Safety Outlier Explorer}
\usage{
safety_outlier_explorer(data, settings, description = "")
}
\arguments{
\item{data}{labs data}

\item{settings}{named list of settings with the following parameters}

\item{description}{description to be included as a subtitle}
}
\value{
returns a chart object
}
\description{
Safety Outlier Explorer
}
\examples{

}
