% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_events.R
\name{stack_events}
\alias{stack_events}
\title{Combine Event Domains}
\usage{
stack_events(data, settings, domains = c("aes", "cm", "ex"))
}
\arguments{
\item{data}{safetyGraphics data object}

\item{settings}{safetyGraphics settings object}

\item{domains}{list of safetyGraphics domains. Adverse events ('aes'), concominate medications ('cm') and exposure ('ex') included by default.}
}
\value{
combined dataset with stacked AE and CM data
}
\description{
Combine Event Domains
}
\examples{
stack_events()

}
